modifs

+++ web/notification/suppr_message.php
@@ -12,11 +12,15 @@ require_once(INC.'/droits.inc.php');
 if(isset($_GET['id']))
 {
 	$db = new DB_connection();
-	$id=$db->quote($_GET['id']);
+	$id = $db->quote($_GET['id']);
 
-	$req="DELETE from Message where id_message='".$id."'";
-
-	$db->DB_query($req);
+	$query = 'SELECT id_message, objet, message, jma, lu FROM Message WHERE utilisateur = 0 AND id_parent = '.$_SESSION['id_parent'].' AND id_message = "'.$id.'"';
+	$db->DB_query($query);
+	if($db->DB_count() > 0)
+	{
+		$req = "DELETE from Message where id_message = '".$id."'";
+		$db->DB_query($req);
+	}
 }
 
 $link = $_SERVER["QUERY_STRING"];
